+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 > This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2025-08-29
+### Added
+* Support for saving images to file as an LDR in PNG format.
+* Support for copying the current edited image into the clipboard [(Feature request)](https://github.com/mcrescas/veriv/issues/8).
+
 ## [1.1.0] - 2025-06-19
 ### Added
 * Now the extension should work both in desktop and web modes.